move getTerminalStatusDate to submissions service
