+++ core/src/test/scala/org/broadinstitute/dsde/rawls/workspace/WorkspaceServiceSpec.scala
@@ -1743,123 +1743,6 @@ class WorkspaceServiceSpec
       }
     }
 
-  "extractOperationIdsFromCromwellMetadata" should "parse workflow metadata" in {
-    val jsonString =
-      """{
-        |  "calls": {
-        |    "hello_and_goodbye.goodbye": [
-        |      {
-        |        "attempt": 1,
-        |        "backendLogs": {
-        |          "log": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-goodbye/goodbye.log"
-        |        },
-        |        "backendStatus": "Success",
-        |        "end": "2019-04-24T13:57:48.998Z",
-        |        "executionStatus": "Done",
-        |        "jobId": "operations/EN2siP2kLRinu-Wt-4-bqRQgw8Sszq0dKg9wcm9kdWN0aW9uUXVldWU",
-        |        "shardIndex": -1,
-        |        "start": "2019-04-24T13:56:22.387Z",
-        |        "stderr": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-goodbye/goodbye-stderr.log",
-        |        "stdout": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-goodbye/goodbye-stdout.log"
-        |      }
-        |    ],
-        |    "hello_and_goodbye.hello": [
-        |      {
-        |        "attempt": 1,
-        |        "backendLogs": {
-        |          "log": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-hello/hello.log"
-        |        },
-        |        "backendStatus": "Success",
-        |        "end": "2019-04-24T13:58:21.978Z",
-        |        "executionStatus": "Done",
-        |        "jobId": "operations/EKCsiP2kLRiu0qj_qdLFq8wBIMPErM6tHSoPcHJvZHVjdGlvblF1ZXVl",
-        |        "shardIndex": -1,
-        |        "start": "2019-04-24T13:56:22.387Z",
-        |        "stderr": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-hello/hello-stderr.log",
-        |        "stdout": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/call-main_workflow/sub.main_workflow/1cf452d0-f18c-4945-aaf4-779402e7b2aa/call-hello_and_goodbye/sub.hello_and_goodbye/0d6768b7-73b3-41c4-b292-de743657c5db/call-hello/hello-stdout.log"
-        |      }
-        |    ]
-        |  },
-        |  "end": "2019-04-24T13:58:23.868Z",
-        |  "id": "0d6768b7-73b3-41c4-b292-de743657c5db",
-        |  "start": "2019-04-24T13:56:20.348Z",
-        |  "status": "Succeeded",
-        |  "workflowName": "sub.hello_and_goodbye",
-        |  "workflowRoot": "gs://fc-2d8ada07-750f-4db8-88ab-307099d54a31/d25c4529-c247-41e0-99fb-1b8fade199d5/most_main_workflow/ccc3fdbe-3cf4-40cf-8a01-4ae77a5d3e5f/"
-        |}""".stripMargin
-
-    import spray.json._
-    val metadataJson = jsonString.parseJson.asJsObject
-    SubmissionsService.extractOperationIdsFromCromwellMetadata(metadataJson) should contain theSameElementsAs Seq(
-      "operations/EN2siP2kLRinu-Wt-4-bqRQgw8Sszq0dKg9wcm9kdWN0aW9uUXVldWU",
-      "operations/EKCsiP2kLRiu0qj_qdLFq8wBIMPErM6tHSoPcHJvZHVjdGlvblF1ZXVl"
-    )
-  }
-
-  behavior of "getTerminalStatusDate"
-
-  // test getTerminalStatusDate
-  private val workflowFinishingTomorrow =
-    testData.submissionMixed.workflows.head.copy(statusLastChangedDate = testDate.plusDays(1))
-  private val submissionMixedDates = testData.submissionMixed.copy(
-    workflows = workflowFinishingTomorrow +: testData.submissionMixed.workflows.tail
-  )
-  private val getTerminalStatusDateTests = Table(
-    ("description", "submission", "workflowId", "expectedOutput"),
-    ("submission containing one completed workflow, no workflowId input",
-     testData.submissionSuccessful1,
-     None,
-     Option(testDate)
-    ),
-    ("submission containing one completed workflow, with workflowId input",
-     testData.submissionSuccessful1,
-     testData.submissionSuccessful1.workflows.head.workflowId,
-     Option(testDate)
-    ),
-    ("submission containing one completed workflow, with nonexistent workflowId input",
-     testData.submissionSuccessful1,
-     Option("thisWorkflowIdDoesNotExist"),
-     None
-    ),
-    ("submission containing several workflows with one finishing tomorrow, no workflowId input",
-     submissionMixedDates,
-     None,
-     Option(testDate.plusDays(1))
-    ),
-    ("submission containing several workflows, with workflowId input for workflow finishing tomorrow",
-     submissionMixedDates,
-     submissionMixedDates.workflows.head.workflowId,
-     Option(testDate.plusDays(1))
-    ),
-    ("submission containing several workflows, with workflowId input for workflow finishing today",
-     submissionMixedDates,
-     submissionMixedDates.workflows(2).workflowId,
-     Option(testDate)
-    ),
-    ("submission containing several workflows, with workflowId input for workflow not finished",
-     submissionMixedDates,
-     submissionMixedDates.workflows.last.workflowId,
-     None
-    ),
-    ("aborted submission, no workflowId input", testData.submissionAborted1, None, Option(testDate)),
-    ("aborted submission, with workflowId input",
-     testData.submissionAborted1,
-     testData.submissionAborted1.workflows.head.workflowId,
-     Option(testDate)
-    ),
-    ("in progress submission, no workflowId input", testData.submissionSubmitted, None, None),
-    ("in progress submission, with workflowId input",
-     testData.submissionSubmitted,
-     testData.submissionSubmitted.workflows.head.workflowId,
-     None
-    )
-  )
-  forAll(getTerminalStatusDateTests) { (description, submission, workflowId, expectedOutput) =>
-    it should s"run getTerminalStatusDate test for $description" in {
-      assertResult(WorkspaceService.getTerminalStatusDate(submission, workflowId))(expectedOutput)
-    }
-  }
-
   behavior of "RequesterPays"
 
   it should "return Unit when adding linked service accounts to workspace" in withTestDataServices { services =>
